Update grid flex properties & Add clear button
The grid now always takes up fixed amount of space on screen, and each
grid-cells that gets added also get smaller to fit-in.
Also added logic to clear grid, or clear all drawing on a press of a
button.
Also updated index.html & stlyes.css to improve the webpage's design.